minor fixes, README populated
