+++ .gitignore
@@ -161,3 +161,5 @@ cython_debug/
 
 # temp dir 
 temp/
+
+output/
